+++ packages/sources/react/src/components/actions/VtmnLink/VtmnLink.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import '@vtmn/css-link/dist/index-with-vars.css';
 import clsx from 'clsx';
 import { VtmnLinkSize } from './types';
-import { objectValuesToString } from '../../../utils/object';
 
 export interface VtmnLinkProps extends React.ComponentPropsWithoutRef<'a'> {
   /**
@@ -49,7 +48,7 @@ export const VtmnLink = ({
         { 'vtmn-link--icon-along': standalone && iconAlong },
         className,
       )}
-      {...objectValuesToString(props)}
+      {...props}
     >
       {children}
     </a>