+++ packages/sources/react/src/utils/object.ts
@@ -1,2 +1,0 @@
-export const objectValuesToString = (object: object) =>
-  Object.keys(object).map((key) => object[key].toString());